Updated docker to latest install revision

+++ system-setup/linux/setup-docker.sh
@@ -3,28 +3,28 @@
 # File:     setup-docker.sh
 #
 # Author:   Cashiuus
-# Created:  09-Mar-2017     Revised: 20-Oct-2020
+# Created:  09-Mar-2017     Revised: 16-Jan-2022
 #
 #-[ Info ]-------------------------------------------------------------------------------
-# Purpose:  Install and configure Docker-CE, the newer version of docker
-#			using official repositories. After install, adds user to 
-#			docker group, starts service, and sets it to autostart.
+# Purpose:  Install and configure Docker
+#     using official repositories. After install, adds user to
+#     docker group, starts service, and sets it to autostart.
 #
 #
 #-[ Notes ]-------------------------------------------------------------------------------
 #
 #
+# - Docker Docs:
+#     - https://docs.docker.com/engine/install/debian/
+#     - https://docs.docker.com/compose/install/
 #
-#
-#-[ Links/Credit ]------------------------------------------------------------------------
-#
-# - Docker Docs: https://docs.docker.com/engine/installation/linux/debian/#os-requirements
-# - Troubleshooting: https://docs.docker.com/engine/installation/linux/linux-postinstall/#kernel-compatibility
+# - Troubleshooting:
+#     - https://docs.docker.com/engine/installation/linux/linux-postinstall/#kernel-compatibility
 #
 #-[ Copyright ]---------------------------------------------------------------------------
 #   MIT License ~ http://opensource.org/licenses/MIT
 ## =======================================================================================
-__version__="0.2.1"
+__version__="0.2.2"
 __author__="Cashiuus"
 ## =============[ CONSTANTS ]============== ##
 START_TIME=$(date +%s)
@@ -102,53 +102,68 @@ function is_process_alive() {
 
 $SUDO apt-get -qq update
 $SUDO apt-get remove --purge docker
+# Full list of apps to remove: docker docker-engine docker.io containerd runc
 # This one may fail so running them separately
 $SUDO apt-get remove --purge docker-engine
 
 # Contents of previous installations may be in /var/lib/docker/
 
 
-
-# Enable the backports repository
-#echo -e "${GREEN}[*]${RESET} Creating backports repo file in /sources.list.d/"
-#file=/etc/apt/sources.list.d/backports.list
-#[[ ! -s "${file}" ]]; then
-#	$SUDO sh -c "echo deb http://httpredir.debian.org/debian jessie-backports main contrib non-free > ${file}"
-#fi
-
 # ========[  Setup the Docker repository ]========== #
 export DEBIAN_FRONTEND=noninteractive
 $SUDO apt-get -qq update
 $SUDO apt-get -y install apt-transport-https ca-certificates \
-  curl gnupg-agent jq software-properties-common
-
-#TODO: Does this work?
-#curl -fsSL https://download.docker.com/linux/debian/gpg | $SUDO apt-key add -
-curl -fsSL https://download.docker.com/linux/$(. /etc/os-release; echo "$ID")/gpg | $SUDO apt-key add -
-
-# Key should be: 9DC8 5822 9FC7 DD38 854A E2D8 8D81 803C 0EBF CD88
-# Verify via: sudo apt-key fingerprint
+  curl gnupg gnupg-agent jq lsb-release software-properties-common
 
 
-# Set up the stable repository. You always need the stable repository, 
-# even if you want to install builds from the edge or test repositories 
+
+# Set up the stable repository. You always need the stable repository,
+# even if you want to install builds from the edge or test repositories
 # as well. To add the edge or test repository, add the word edge or test
 # (or both) after the word stable in the commands below.
-echo -e "${GREEN}[*]${RESET} Adding docker repository"
-$SUDO add-apt-repository \
+
+OS=$(. /etc/os-release; echo "$ID")
+
+if [[ $OS == "kali" ]];then
+  #curl -fsSL https://download.docker.com/linux/debian/gpg | $SUDO apt-key add -
+  curl -fsSL https://download.docker.com/linux/debian/gpg | $SUDO gpg --dearmor -o /usr/share/keyrings/docker-archive-keyring.gpg
+  $SUDO chmod a+r /usr/share/keyrings/docker-archive-keyring.gpg
+  echo -e "${GREEN}[*]${RESET} Adding docker repository"
+  #$SUDO add-apt-repository \
+  "deb [arch=amd64] https://download.docker.com/linux/$(. /etc/os-release; echo "$ID_LIKE") \
+  buster \
+  stable"
+  echo \
+  "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/docker-archive-keyring.gpg] https://download.docker.com/linux/debian \
+  buster stable" | $SUDO tee /etc/apt/sources.list.d/docker.list > /dev/null
+else
+  #curl -fsSL https://download.docker.com/linux/$(. /etc/os-release; echo "$ID")/gpg | $SUDO apt-key add -
+  curl -fsSL https://download.docker.com/linux/$(. /etc/os-release; echo "$ID")/gpg | $SUDO gpg --dearmor -o /usr/share/keyrings/docker-archive-keyring.gpg
+  $SUDO chmod a+r /usr/share/keyrings/docker-archive-keyring.gpg
+  echo -e "${GREEN}[*]${RESET} Adding docker repository"
+  #$SUDO add-apt-repository \
   "deb [arch=amd64] https://download.docker.com/linux/$(. /etc/os-release; echo "$ID") \
   $(lsb_release -cs) \
   stable"
-# The above will output as "https://download.docker.com/linux/debian buster stable"
+  echo \
+  "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/docker-archive-keyring.gpg] https://download.docker.com/linux/debian \
+  $(lsb_release -cs) stable" | $SUDO tee /etc/apt/sources.list.d/docker.list > /dev/null
+fi
+
+# Verify key via: sudo apt-key fingerprint
+
+if [[ ! -f /usr/share/keyrings/docker-archive-keyring.gpg ]]; then
+  echo -e "[ERR] Failed to setup docker gpg keyring. Check and try again!"
+  exit 1
+fi
 
 
+### Install Docker Engine
 
 # If you don't specify a version, it will install the latest release
-# On production systems, you should install a specific version of Docker
-# instead of defaulting to the latest.
-#   List available versions: 	apt-cache madison docker-ce
-#   Better: 					$(apt-cache madison docker-ce | cut -d '|' -f2)
-#   Specify Install Version: 	"sudo apt-get -y install docker-ce=18.06.1~ce~3-0~debian"
+#   List available versions:  apt-cache madison docker-ce
+#   Better:           $(apt-cache madison docker-ce | cut -d '|' -f2)
+#   Specify Install Version:  "sudo apt-get -y install docker-ce=18.06.1~ce~3-0~debian"
 echo -e "${GREEN}[*]${RESET} Installing Docker-CE via apt-get"
 $SUDO apt-get -qq update
 $SUDO apt-get -y install docker-ce docker-ce-cli containerd.io
@@ -156,7 +171,7 @@ $SUDO apt-get -y install docker-ce docker-ce-cli containerd.io
 
 # How to list available versions if you require a specific docker version on a production system
 #apt-cache madison docker-ce
-#	Output: docker-ce | 17.06.0~ce-0~debian | https://download.docker.com/linux/debian jessie/stable amd64 Packages
+# Output: docker-ce | 17.06.0~ce-0~debian | https://download.docker.com/linux/debian jessie/stable amd64 Packages
 # Install a specific version
 #sudo apt-get install docker-ce=<VERSION_STRING>
 
@@ -167,54 +182,70 @@ $SUDO apt-get -y install docker-ce docker-ce-cli containerd.io
 
 # Download the release binary from their Github releases
 # Site: https://github.com/docker/compose/releases
-#$SUDO curl -L https://github.com/docker/compose/releases/download/1.16.1/docker-compose-`uname -s`-`uname -m` -o /usr/local/bin/docker-compose
-
-$SUDO curl -L "https://github.com/docker/compose/releases/download/1.27.4/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
-
-# Make it executable
+$SUDO curl -L "https://github.com/docker/compose/releases/download/1.29.2/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
 $SUDO chmod +x /usr/local/bin/docker-compose
-
+# If /usr/local/bin isn't typically in your path, you can symlink it
+#$SUDO ln -s /usr/local/bin/docker-compose /usr/bin/docker-compose
 
 # Setup bash/zsh completions
 # Place the completion script in /etc/bash_completion.d/
-$SUDO curl -L https://raw.githubusercontent.com/docker/compose/1.27.4/contrib/completion/bash/docker-compose -o /etc/bash_completion.d/docker-compose
+echo -e "[*] Your current shell is $SHELL_NAME. Adding docker completions for it now."
+if [[ "$SHELL_NAME" == "bash" ]]; then
+  $SUDO curl -L https://raw.githubusercontent.com/docker/compose/1.29.2/contrib/completion/bash/docker-compose -o /etc/bash_completion.d/docker-compose
+elif [[ "$SHELL_NAME" == "zsh" ]]; then
+  mkdir -p ~/.zsh/completion > /dev/null
+  curl \
+    -L https://raw.githubusercontent.com/docker/compose/1.29.2/contrib/completion/zsh/_docker-compose \
+    -o ~/.zsh/completion/_docker-compose
+  # Add this dir to your fpath
+  grep -q '^fpath=' "${SHELL_FILE}" 2>/dev/null \
+    || echo 'fpath=(~/.zsh/completion $fpath)' >> "${SHELL_FILE}"
+  grep -q '^autoload -Uz' "${SHELL_FILE}" 2>/dev/null \
+    || echo 'autoload -Uz compinit && compinit -i' >> "${SHELL_FILE}"
+  # reload your shell
+  exec $SHELL -l
+fi
+
+
 
 
 
 
 
 ### =========[ Manage Docker as a Non-Root User ]======== ###
-# Ref: https://docs.docker.com/engine/installation/linux/linux-postinstall/#manage-docker-as-a-non-root-user
+# Ref: https://docs.docker.com/engine/install/linux-postinstall/#manage-docker-as-a-non-root-user
 # You must do these steps on a Debian system, or docker commands will fail due to permissions
 
 # Create group 'docker' -- may already exist
 echo -e "${GREEN}[*]${RESET} Configuring user ${ORANGE}${USER}${RESET} into the docker group"
 $SUDO groupadd docker
 
-# Add user to group 'docker'
+# Add current user to group 'docker'
 $SUDO usermod -aG docker $USER
+newgrp docker
 
 echo -e "${GREEN}[*]${RESET} User has been added to to the docker group"
 echo -e "${GREEN}[*]${RESET} Logout and back in. Then, you should be able to run 'docker run hello-world' and 'docker info' without sudo"
-pause
+sleep 5s
 
 # If you ran 'sudo docker run ...' before adding non-root capability, you
 # will now need to fix the default ~/.docker/ directory permissions
 if [[ $EUID -ne 0 ]]; then
-	if [[ -d "$HOME/.docker" ]]; then
-		$SUDO chown "$USER":"$USER" /home/"$USER"/.docker -R
-		$SUDO chmod g+rwx "$HOME/.docker" -R
-	fi
+  if [[ -d "$HOME/.docker" ]]; then
+    $SUDO chown "$USER":"$USER" /home/"$USER"/.docker -R
+    $SUDO chmod g+rwx "$HOME/.docker" -R
+  fi
 else
-	echo -e "${YELLOW}[WARN]${RESET} You are already root, skipping permissions changes"
+  echo -e "${YELLOW}[WARN]${RESET} You are already root, skipping permissions changes"
 fi
 
 # Start the service
 echo -e "${GREEN}[*]${RESET} Starting Docker daemon service"
-$SUDO systemctl start docker
+$SUDO systemctl start docker.service
 # Set it for autostart
 echo -e "${GREEN}[*]${RESET} Setting Docker service to autostart on boot"
-$SUDO systemctl enable docker
+$SUDO systemctl enable docker.service
+#$SUDO systemctl enable containerd.service
 
 # Default Ubuntu container for testing
 #docker run -it ubuntu bash
@@ -224,8 +255,7 @@ $SUDO systemctl enable docker
 
 
 # Verify it's working
-echo -e "\n\n${GREEN}[*]${RESET} Verifying install. You should see the hello-world docker run below, and without requiring sudo to do it..."
-docker run hello-world
+echo -e "\n\n${GREEN}[*]${RESET} Docker installation complete. After script ends, verify it was successful by logging out and back in for group to take effect. Then, run: docker run hello-world."
 
 
 function finish() {
@@ -247,7 +277,12 @@ trap finish EXIT
 ## ===================================================================================== ##
 ## =========================[ File Code Help :: Core Notes ]============================ ##
 #
-
+# Enable the backports repository
+#echo -e "${GREEN}[*]${RESET} Creating backports repo file in /sources.list.d/"
+#file=/etc/apt/sources.list.d/backports.list
+#[[ ! -s "${file}" ]]; then
+# $SUDO sh -c "echo deb http://httpredir.debian.org/debian jessie-backports main contrib non-free > ${file}"
+#fi
 ## =========[    ]========= ##
 
 ## ==================================================================================== ##
